requirements.txt
# Web Framework
fastapi==0.115.6

# ASGI Server
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0

# HTTP Client
httpx==0.28.1

# Data Validation
pydantic==2.10.4
pydantic-settings==2.7.0

# Environment variables
python-dotenv==1.0.1

# Code Quality
ruff==0.8.5
